backend/modules/plagiarism/semanticsimiliarty/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-multipart>=0.0.6

# HuggingFace & Deep Learning (for fine-tuning)
datasets>=2.14.0
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0
huggingface_hub>=0.20.0

# Vector Search & Similarity
faiss-cpu>=1.7.4

# Visualization
matplotlib>=3.7.0

# PDF/DOCX parsing
PyPDF2>=3.0.0
python-docx>=0.8.11

# Environment variables
python-dotenv>=1.0.0

# Database
mysql-connector-python>=8.0.0

# Authentication
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
google-auth>=2.0.0
email-validator>=2.0.0

# Enhanced Web Scraping (FREE alternative to Google API)
duckduckgo-search>=8.0.0
trafilatura>=2.0.0
playwright>=1.40.0
